+++ src/compojure/api/meta.clj
@@ -9,7 +9,6 @@
             [ring.swagger.common :as rsc]
             [ring.swagger.json-schema :as js]
             [ring.util.http-response :refer [internal-server-error]]
-            [slingshot.slingshot :refer [throw+]]
             [schema.core :as s]
             [schema.coerce :as sc]
             [schema.utils :as su]
@@ -65,7 +64,8 @@
           (let [coerce (coercer (rsc/value-of schema) matcher)
                 body (coerce (:body response))]
             (if (su/error? body)
-              (throw+ (assoc body :type ::ex/response-validation))
+              (throw (ex-info "Response validation error"
+                              (assoc body :type ::ex/response-validation)))
               (assoc response
                 ::serializable? true
                 :body body)))
@@ -78,7 +78,7 @@
       (let [coerce (coercer schema matcher)
             result (coerce value)]
         (if (su/error? result)
-          (throw+ (assoc result :type ::ex/request-validation))
+          (throw (ex-info "Request validation failed" (assoc result :type ::ex/request-validation)))
           result))
       value)))
 
@@ -153,7 +153,7 @@
 (defmethod restructure-param :tags [_ tags acc]
   (update-in acc [:parameters :tags] (comp set into) tags))
 
-; Defines a return type and coerced the return value of a body against it.
+; Defines a return type and coerces the return value of a body against it.
 ; Examples:
 ; :return MySchema
 ; :return {:value String}