+++ src/compojure/api/middleware.clj
@@ -11,7 +11,6 @@
             [ring.swagger.middleware :as rsm]
             [ring.swagger.coerce :as coerce]
             [ring.util.http-response :refer :all]
-            [slingshot.slingshot :refer [try+ throw+]]
             [schema.core :as s])
   (:import [com.fasterxml.jackson.core JsonParseException]
            [org.yaml.snakeyaml.parser ParserException]
@@ -23,15 +22,15 @@
 
 (def rethrow-exceptions? ::rethrow-exceptions?)
 
-(defn- call-error-handler [error-handler error error-type request]
+(defn- call-error-handler [error-handler error data request]
   (try
-    (error-handler error error-type request)
-    (catch ArityException e
+    (error-handler error data request)
+    (catch ArityException _
       (println "WARNING: Error-handler arity has been changed.")
       (error-handler error))))
 
 (defn wrap-exceptions
-  "Catches all exceptions and delegates to right error handler accoring to :type of Exceptions
+  "Catches all exceptions and delegates to correct error handler according to :type of Exceptions
    - **:handlers** - a map from exception type to handler
      - **:compojure.api.exception/default** - Handler used when exception type doesn't match other handler,
                                               by default prints stack trace."
@@ -39,18 +38,16 @@
   (let [default-handler (get handlers ::ex/default ex/safe-handler)]
     (assert (fn? default-handler) "Default exception handler must be a function.")
     (fn [request]
-      (try+
+      (try
         (handler request)
-        (catch (get % :type) {:keys [type] :as data}
-          (let [type (or (get ex/legacy-exception-types type) type)]
-            (if-let [handler (get handlers type)]
-              (call-error-handler handler (:throwable &throw-context) data request)
-              (call-error-handler default-handler (:throwable &throw-context) data request))))
-        (catch Object _
-          ; FIXME: Used for validate
-          (if (rethrow-exceptions? request)
-            (throw+)
-            (call-error-handler default-handler (:throwable &throw-context) nil request)))))))
+        (catch Throwable e
+          (let [{:keys [type] :as data} (ex-data e)
+                type (or (get ex/legacy-exception-types type) type)
+                handler (or (get handlers type) default-handler)]
+            ; FIXME: Used for validate
+            (if (rethrow-exceptions? request)
+              (throw e)
+              (call-error-handler handler e data request))))))))
 
 ;;
 ;; Component integration
@@ -128,11 +125,11 @@
   ;; i.e. (handler req) is inside try-catch. If r-m-f was changed to catch only
   ;; exceptions from parsing the request, we wouldn't need to check the exception class.
   (if (or (instance? JsonParseException e) (instance? ParserException e))
-    (throw+ {:type ::ex/request-parsing} e)
-    (throw+ e)))
+    (throw (ex-info "Error parsing request" {:type ::ex/request-parsing} e))
+    (throw e)))
 
 (defn serializable?
-  "Predicate which return true if the response body is serializable.
+  "Predicate which returns true if the response body is serializable.
    That is, return type is set by :return compojure-api key or it's
    a collection."
   [_ {:keys [body] :as response}]
@@ -168,9 +165,12 @@
                                        :compojure.api.exception/response-validation compojure.api.exception/response-validation-handler
                                        :compojure.api.exception/default             compojure.api.exception/safe-handler}
 
+                                      Note: Because the handlers are merged into default handlers map, to disable default handler you
+                                      need to provide `nil` value as handler.
+
                                       Note: To catch Schema errors use {:schema.core/error compojure.api.exception/schema-error-handler}
 
-                                      Note: Adding alias for exception namespace makes it easier to define these options.
+                                      Note: Adding an alias for exception namespace makes it easier to define these options.
 
    - **:format**                    for ring-middleware-format middlewares
        - **:formats**                 sequence of supported formats, e.g. `[:json-kw :edn]`
@@ -183,7 +183,7 @@
                                     e.g. `{:ignore-missing-mappings? true}`
 
    - **:coercion**                  A function from request->type->coercion-matcher, used
-                                    in enpoint coersion for :json, :query and :response.
+                                    in endpoint coercion for :json, :query and :response.
                                     Defaults to `compojure.api.middleware/default-coercion-matchers`
 
    - **:components**                Components which should be accessible to handlers using